+++ pytsite/google/_maps/__init__.py
@@ -2,7 +2,7 @@
 """
 # Public API
 from . import _widget as widget, _geocoding as geocoding
-from ._api import code, decode, get_map_link
+from ._api import encode, decode, get_map_link
 
 __author__ = 'Alexander Shepetko'
 __email__ = 'a@shepetko.com'